fix(engine): place skipped times forward, keep far dates, rescale times exactly
The branch review found that a wall-clock time a zone east of UTC skips
moved backward (the reference shared the rule), that a Date64 beyond a
Date32's days became null in a timestamp column and was refused on its
way to JSON, and that Arrow widens times of day outside a day unchecked.
Skipped times take the offset in force a day earlier, found by bisection
in the reference; dates go to timestamps from their own days and to JSON
as Date64s, rendered as dates; times are rescaled with checked
arithmetic. The differential now draws far Date64s, times outside a day,
eastern zones and the clocks zones shifted, and fails if any fix is
reverted. Text renders into a reused buffer.

+++ crates/rdlt-engine/src/drawn.rs
@@ -23,7 +23,7 @@ pub(crate) enum Scalar {
     Utf8(String),
     Binary(Vec<u8>),
     /// Days since the epoch.
-    Date(i32),
+    Date(i64),
     /// A time of day, a timestamp or a duration, in its type's unit.
     Temporal(i64),
     Uuid([u8; 16]),
